+++ sdk/src/org.graalvm.polyglot/src/org/graalvm/polyglot/Value.java
@@ -42,6 +42,13 @@ package org.graalvm.polyglot;
 
 import java.math.BigDecimal;
 import java.math.BigInteger;
+import java.time.Duration;
+import java.time.Instant;
+import java.time.LocalDate;
+import java.time.LocalDateTime;
+import java.time.LocalTime;
+import java.time.ZoneId;
+import java.time.ZonedDateTime;
 import java.util.List;
 import java.util.Map;
 import java.util.Objects;
@@ -57,7 +64,7 @@ import org.graalvm.polyglot.proxy.Proxy;
  * values are bound to a {@link Context context}. If the context is closed then all value operations
  * throw an {@link IllegalStateException}.
  * <p>
- * Polyglot values have one of the following types:
+ * Polyglot values have one of the following type combinations:
  * <ul>
  * <li>{@link #isNull() Null}: This value represents a <code>null</code> like value. Certain
  * languages might use a different name or use multiple values to represent <code>null</code> like
@@ -69,6 +76,11 @@ import org.graalvm.polyglot.proxy.Proxy;
  * accessed using {@link #asBoolean()}.
  * <li>{@link #isString() String}: This value represents a string value. The string value can be
  * accessed using {@link #asString()}.
+ * <li>{@link #isDate() Date}, {@link #isTime() Time} or {@link #isTimeZone() Timezone}: This value
+ * represents a date, time or timezone. Multiple types may return <code>true</code> at the same
+ * time.
+ * <li>{@link #isDuration() Duration}: This value represents a duration value. The duration value
+ * can be accessed using {@link #asDuration()}.
  * <li>{@link #isHostObject() Host Object}: This value represents a value of the host language
  * (Java). The original Java value can be accessed using {@link #asHostObject()}.
  * <li>{@link #isProxyObject() Proxy Object}: This value represents a {@link Proxy proxy} value.
@@ -98,6 +110,22 @@ import org.graalvm.polyglot.proxy.Proxy;
  * <p>
  * Polyglot values may be converted to host objects using {@link #as(Class)}. In addition values may
  * be created from Java values using {@link Context#asValue(Object)}.
+ *
+ * <h3>Naive and aware dates and times</h3>
+ * <p>
+ * If a date or time value has a {@link #isTimeZone() timezone} then it is called <i>aware</i>,
+ * otherwise <i>naive</i>.
+ * <p>
+ * An aware time and date has sufficient knowledge of applicable algorithmic and political time
+ * adjustments, such as time zone and daylight saving time information, to locate itself relative to
+ * other aware objects. An aware object is used to represent a specific moment in time that is not
+ * open to interpretation.
+ * <p>
+ * A naive time and date does not contain enough information to unambiguously locate itself relative
+ * to other date/time objects. Whether a naive object represents Coordinated Universal Time (UTC),
+ * local time, or time in some other timezone is purely up to the program, just like it is up to the
+ * program whether a particular number represents metres, miles, or mass. Naive objects are easy to
+ * understand and to work with, at the cost of ignoring some aspects of reality.
  *
  * @see Context
  * @see Engine
@@ -770,6 +798,20 @@ public final class Value {
      * boolean}. Primitive {@link Boolean boolean.class} literal is also supported. The primitive
      * class literal throws a {@link NullPointerException} if the value represents {@link #isNull()
      * null}.
+     * <li><code>{@link LocalDate}.class</code> is supported if the value is a {@link #isDate()
+     * date}</li>
+     * <li><code>{@link LocalTime}.class</code> is supported if the value is a {@link #isTime()
+     * time}</li>
+     * <li><code>{@link LocalDateTime}.class</code> is supported if the value is a {@link #isDate()
+     * date} and {@link #isTime() time}.</li>
+     * <li><code>{@link Instant}.class</code> is supported if the value is an {@link #isInstant()
+     * instant}.</li>
+     * <li><code>{@link ZonedDateTime}.class</code> is supported if the value is a {@link #isDate()
+     * date}, {@link #isTime() time} and {@link #isTimeZone() timezone}.</li>
+     * <li><code>{@link ZoneId}.class</code> is supported if the value is a {@link #isTimeZone()
+     * timezone}.</li>
+     * <li><code>{@link Duration}.class</code> is supported if the value is a {@link #isDuration()
+     * duration}.</li>
      * <li>Any Java type in the type hierarchy of a {@link #isHostObject() host object}.
      * <li><code>{@link Object}.class</code> is always supported. See section Object mapping rules.
      * <li><code>{@link Map}.class</code> is supported if the value has {@link #hasMembers()
@@ -816,7 +858,7 @@ public final class Value {
      * <b>JavaScript Usage Examples:</b>
      *
      * <pre>
-     * Context context = Context.create();
+     * Context context = Context.newBuilder().allowHostAccess(HostAccess.ALL).build();
      * assert context.eval("js", "undefined").as(Object.class) == null;
      * assert context.eval("js", "'foobar'").as(String.class).equals("foobar");
      * assert context.eval("js", "42").as(Integer.class) == 42;
@@ -971,6 +1013,181 @@ public final class Value {
      */
     public SourceSection getSourceLocation() {
         return impl.getSourceLocation(receiver);
+    }
+
+    /**
+     * Returns <code>true</code> if this object represents a date, else <code>false</code>. If the
+     * receiver is also a {@link #isTimeZone() timezone} then the date is aware, otherwise it is
+     * naive.
+     *
+     * @throws ClassCastException if polyglot value could not be mapped to the target type.
+     * @throws NullPointerException if the target type is null.
+     * @throws PolyglotException if the conversion triggered a guest language error.
+     * @throws IllegalStateException if the underlying context is already closed.
+     * @see #asDate()
+     * @since 20.0.0 beta 2
+     */
+    public boolean isDate() {
+        return impl.isDate(receiver);
+    }
+
+    /**
+     * Returns the receiver as date if this object represents a {@link #isDate() date}. The returned
+     * date is either aware if the receiver has a {@link #isTimeZone() timezone} otherwise it is
+     * naive.
+     *
+     * @throws ClassCastException if polyglot value could not be mapped to the target type.
+     * @throws NullPointerException if the target type is null.
+     * @throws PolyglotException if the conversion triggered a guest language error.
+     * @throws IllegalStateException if the underlying context is already closed.
+     * @see #isDate()
+     * @since 20.0.0 beta 2
+     */
+    public LocalDate asDate() {
+        return impl.asDate(receiver);
+    }
+
+    /**
+     * Returns <code>true</code> if this object represents a time, else <code>false</code>. If the
+     * receiver is also a {@link #isTimeZone() timezone} then the time is aware, otherwise it is
+     * naive.
+     *
+     * @throws IllegalStateException if the underlying context is already closed.
+     * @see #asTime()
+     * @since 20.0.0 beta 2
+     */
+    public boolean isTime() {
+        return impl.isTime(receiver);
+    }
+
+    /**
+     * Returns the receiver as time if this object represents a {@link #isTime() time}. The returned
+     * time is either aware if the receiver has a {@link #isTimeZone() timezone} otherwise it is
+     * naive.
+     *
+     * @throws ClassCastException if polyglot value could not be mapped to the target type.
+     * @throws NullPointerException if the target type is null.
+     * @throws PolyglotException if the conversion triggered a guest language error.
+     * @throws IllegalStateException if the underlying context is already closed.
+     * @see #isTime()
+     * @since 20.0.0 beta 2
+     */
+    public LocalTime asTime() {
+        return impl.asTime(receiver);
+    }
+
+    /**
+     * Returns <code>true</code> if the receiver represents an instant. If a value is an instant
+     * then it is also a {@link #isDate() date}, {@link #isTime() time} and {@link #isTimeZone()
+     * timezone}.
+     *
+     * This method is short-hand for:
+     *
+     * <pre>
+     * v.{@linkplain #isDate() isDate}() && v.{@link #isTime() isTime}() && v.{@link #isTimeZone() isTimeZone}()
+     * </pre>
+     *
+     * @throws IllegalStateException if the underlying context is already closed.
+     * @see #isDate()
+     * @see #isTime()
+     * @see #isInstant()
+     * @see #asInstant()
+     * @since 20.0.0 beta 2
+     */
+    public boolean isInstant() {
+        return isDate() && isTime() && isTimeZone();
+    }
+
+    /**
+     * Returns the receiver as instant if this object represents an {@link #isInstant() instant}. If
+     * a value is an instant then it is also a {@link #isDate() date}, {@link #isTime() time} and
+     * {@link #isTimeZone() timezone}. Using this method may be more efficient than reconstructing
+     * the timestamp from the date, time and timezone data.
+     * <p>
+     * The following assertion always holds if {@link #isInstant()} returns <code>true</code>:
+     *
+     * <pre>
+     * ZoneId zone = getTimeZone(receiver);
+     * LocalDate date = getDate(receiver);
+     * LocalTime time = getTime(receiver);
+     * assert ZonedDateTime.of(date, time, zone).toInstant().equals(getInstant(receiver));
+     * </pre>
+     *
+     * @throws ClassCastException if polyglot value could not be mapped to the target type.
+     * @throws NullPointerException if the target type is null.
+     * @throws PolyglotException if the conversion triggered a guest language error.
+     * @throws IllegalStateException if the underlying context is already closed.
+     * @see #isDate()
+     * @see #isTime()
+     * @see #isTimeZone()
+     * @since 20.0.0 beta 2
+     */
+    public Instant asInstant() {
+        return impl.asInstant(receiver);
+    }
+
+    /**
+     * Returns <code>true</code> if this object represents a timezone, else <code>false</code>. The
+     * interpretation of timezone objects may vary:
+     * <ul>
+     * <li>If {@link #isDate()} and {@link #isTime()} return <code>true</code>, then the returned
+     * date or time information is aware of this timezone.
+     * <li>If {@link #isDate()} and {@link #isTime()} returns <code>false</code>, then it represents
+     * just timezone information.
+     * </ul>
+     * Objects with only time or only date information must not have timezone information attached,
+     * as aware date or time information always consist of both date and time. If this rule is
+     * violated then an {@link AssertionError} is thrown if assertions are enabled.
+     * <p>
+     * If this method is implemented then also {@link #asTimeZone()} must be implemented.
+     *
+     * @throws IllegalStateException if the underlying context is already closed.
+     * @see #asTimeZone()
+     * @see #asInstant()
+     * @since 20.0.0 beta 2
+     */
+    public boolean isTimeZone() {
+        return impl.isTimeZone(receiver);
+    }
+
+    /**
+     * Returns the receiver as timestamp if this object represents a {@link #isTimeZone() timezone}.
+     *
+     * @throws ClassCastException if polyglot value could not be mapped to the target type.
+     * @throws PolyglotException if the conversion triggered a guest language error.
+     * @throws IllegalStateException if the underlying context is already closed.
+     * @throws NullPointerException if the target type is null.
+     * @see #isTimeZone()
+     * @since 20.0.0 beta 2
+     */
+    public ZoneId asTimeZone() {
+        return impl.asTimeZone(receiver);
+    }
+
+    /**
+     * Returns <code>true</code> if this object represents a duration, else <code>false</code>.
+     *
+     * @throws IllegalStateException if the underlying context is already closed.
+     * @see Duration
+     * @see #asDate()
+     * @since 20.0.0 beta 2
+     */
+    public boolean isDuration() {
+        return impl.isDuration(receiver);
+    }
+
+    /**
+     * Returns the receiver as duration if this object represents a {@link #isDuration() duration}.
+     *
+     * @throws ClassCastException if polyglot value could not be mapped to the target type.
+     * @throws PolyglotException if the conversion triggered a guest language error.
+     * @throws IllegalStateException if the underlying context is already closed.
+     * @throws NullPointerException if the target type is null.
+     * @see #isDuration()
+     * @since 20.0.0 beta 2
+     */
+    public Duration asDuration() {
+        return impl.asDuration(receiver);
     }
 
     /**